Tweak coding conventions (#747)
After a lot of discussion, I'm backpedaling and making this document what the (main) conventions of the codebase *are* rather than trying to prescribe what I think they should be. In particular:

* Go back to `m_` for implementation-details member names, since that is the main convention in place for them

* Put the `k` prefix back on C-style `enum` types since that is how they are currently being defined.

My inclination is to just stick with what we have for now to avoid extended debates. Its a bit of an inconsistent mess, but its enough for getting work done.